requirements.txt
# MinerU - PDF to Markdown converter
mineru[all]

# Translation
dashscope

# DOCX conversion
python-docx
